Finally... unify syscall_handler for all architectures :)
